requirements.txt
numpy==1.25.2
pandas==2.0.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
